% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/csv_target.R
\docType{data}
\name{CSVtarget}
\alias{CSVtarget}
\title{CSV target specifies an csv file stored on your local drive}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7ffb5b0b7b10> 
 - attr(*, "name")= chr "CSVtarget_generator"
}}
\usage{
CSVtarget
}
\description{
CSV target specifies an csv file stored on your local drive
}
\keyword{datasets}

