% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/local_scheduler.R
\docType{data}
\name{local_scheduler}
\alias{local_scheduler}
\title{Local scheduler
Linearly executes the tasks (if needed)}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fb19c48f1f0> 
 - attr(*, "name")= chr "local_scheduler_generator"
}}
\usage{
local_scheduler
}
\description{
Local scheduler
Linearly executes the tasks (if needed)
}
\keyword{datasets}

