% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/R_target.R
\docType{data}
\name{Rtarget}
\alias{Rtarget}
\title{R target specifies an object with a given name in .ruigi_env}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7ffb5b07b9c0> 
 - attr(*, "name")= chr "CSVtarget_generator"
}}
\usage{
Rtarget
}
\description{
R target specifies an object with a given name in .ruigi_env
}
\keyword{datasets}

