% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ruigi_target.R
\docType{data}
\name{ruigi_target}
\alias{ruigi_target}
\title{A minimal deliverable in ruigi. A target is an interface to an object (file, RAM address,
database table, s3 bucket, etc.), with \code{read} and \code{write} methods.}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fb19c393460> 
 - attr(*, "name")= chr "ruigi_target_generator"
}}
\usage{
ruigi_target
}
\description{
A minimal deliverable in ruigi. A target is an interface to an object (file, RAM address,
database table, s3 bucket, etc.), with \code{read} and \code{write} methods.
}
\keyword{datasets}

