% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ruigi_task.R
\docType{data}
\name{ruigi_task}
\alias{ruigi_task}
\title{Minimal unit of computation in Ruigi. A task is a bundle that contains dependencies,
a target (every node should produce one and only one target), and a function that maps the
dependencies to this target. Ruigi will organize and determine the right order of computation
of multiple nodes formed together into one pipeline.}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7fb19c34db20> 
 - attr(*, "name")= chr "ruigi_task_generator"
}}
\usage{
ruigi_task
}
\description{
Minimal unit of computation in Ruigi. A task is a bundle that contains dependencies,
a target (every node should produce one and only one target), and a function that maps the
dependencies to this target. Ruigi will organize and determine the right order of computation
of multiple nodes formed together into one pipeline.
}
\keyword{datasets}

