% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/s3_target.R
\docType{data}
\name{S3target}
\alias{S3target}
\title{S3 target specifies an object with a given path in your default s3mpi bucket}
\format{\preformatted{Class 'R6ClassGenerator' <environment: 0x7ffb5b1e1a38> 
 - attr(*, "name")= chr "S3target_generator"
}}
\usage{
S3target
}
\description{
S3 target specifies an object with a given path in your default s3mpi bucket
}
\keyword{datasets}

