% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline}
\alias{pipeline}
\title{Define the pipeline and watch it get executed}
\usage{
pipeline(tasks, to, remote = FALSE)
}
\arguments{
\item{tasks}{list. A list of ruigi_tasks.}

\item{to}{ruigi_target. The scheduler will stop executing the graph once this target is created.}

\item{remote}{logical. Choose between a local and remote scheduler.}
}
\description{
Define the pipeline and watch it get executed
}

